Audit pass: consolidate kubectl tokenizer + drop unused barrel exports
Two findings from a sweep over the auto-remediation code:

1) Two near-duplicate kubectl-string tokenizers:
   - parseKubectlCommandString (agent-core/handlers/ops.ts, P2)
   - tokenizeKubectlCommand     (api-gateway/services/ops-command-runner-service.ts)
   Both turn a 'kubectl ...' string into argv. The agent-core version is
   strict (refuses shell metacharacters like $/|/;/&/>/<, refuses
   unterminated quotes); the api-gateway version is lenient.

   Consolidate on the strict one and move it into the adapters package
   alongside parseKubectlArgv (these are sibling concerns: argv->meaning
   and string->argv, both kubectl-shaped). Both call sites now import
   from @agentic-obs/adapters.

   Behavior change for the api-gateway path: a command like
   'kubectl get pods | grep web' now tokenizes to [] -> denied instead
   of ["kubectl", "get", "pods", "|", "grep", "web"] -> kubectl
   exec error. The strict gate matches what classifyOpsCommand already
   intends ('only kubectl commands are supported'), so this is
   defense-in-depth, not a regression.

2) Two unused barrel exports added in #105:
     BackgroundAgentRunInput
     ISaTokenResolver
   Zero outside callers. Drop them — keep runBackgroundAgent +
   BackgroundRunnerDeps which the dispatcher uses.

No file-size god-files were found. PlanExecutorService at 476 LOC is
a single state machine (5 public methods + 6 privates, all about
plan lifecycle); splitting would be artificial.

ops-command-runner-service.ts at 370 LOC has classifyOpsCommand co-
located with the runner class. Pre-existing (not introduced by this
feature), so per CLAUDE.md 'don't refactor things that aren't broken'
left as-is.

Tests: full suite 1474 / 16 skipped (no change). Lint clean.

Refs: option D from the post-feature follow-up.
